refactor: Update calculateHours function in JobCard component
